refactor(member): extract MemberCard from HiringContact
Lift the visual card (avatar, name, role, bio, LinkedIn) out of the
jobs-only HiringContact into a generic MemberCard reusable from blog
and stories. HiringContact becomes a thin wrapper that pins the
"@ Ocobo" context suffix on the role line.
